backend/requirements.txt
numpy
pandas
scikit-learn
tensorflow
torch
scipy
matplotlib
seaborn
jupyter
flask
flask-cors
flask-socketio
cryptography
pycryptodome
optuna
mlflow
psycopg2-binary
redis